Fix Angular Refactor functions structure Add scripts
