+++ apps/fx-angular-app/src/app/services/historical-rates/historical-rates.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { DatesService } from '../dates/dates.service';
 import { BehaviorSubject, combineLatest } from 'rxjs';
 import {
   FxEntries,
@@ -15,36 +14,18 @@ import { CurrencySelectionsService } from '../currency-selections/currency-selec
 })
 export class HistoricalRates {
   private entries = new BehaviorSubject<FxEntries>([]);
-  private filterCache = new Map<string, FxEntries>();
   fxEntries = this.entries.asObservable();
 
-  constructor(
-    private dateService: DatesService,
-    private currencySelection: CurrencySelectionsService
-  ) {
+  constructor(private currencySelection: CurrencySelectionsService) {
     combineLatest([
-      this.dateService.dates,
       this.currencySelection.base,
       this.currencySelection.quote
-    ]).subscribe(async ([dates, base, quote]) => {
-      const filterCacheKey = toFilterCacheKey({
-        dates,
-        base,
-        quote
-      });
-      if (!this.filterCache.has(filterCacheKey)) {
-        const unfilteredRates = await AlphaVantageClient.getHistoricalRates(
-          base,
-          quote
-        );
-        this.filterCache.set(
-          filterCacheKey,
-          enttriesFromHistoricalServerResponse(unfilteredRates).filter(
-            filterFromDates(dates)
-          )
-        );
-      }
-      this.entries.next(this.filterCache.get(filterCacheKey));
+    ]).subscribe(async ([base, quote]) => {
+      this.entries.next(
+        enttriesFromHistoricalServerResponse(
+          await AlphaVantageClient.getHistoricalRates(base, quote)
+        )
+      );
     });
   }
 }
