+++ apps/fx-react-app/src/app/components/historical-rates-chart/historical-rates-chart.tsx
@@ -1,14 +1,12 @@
 import React, { useMemo } from 'react';
 import { HighchartsWrapper } from '../highcharts-wrapper/highcharts-wrapper';
 import {
+  historicalChartOptions,
+  fxEntriesToXpointData,
+  lineChartOptions,
   FxEntries,
   CurrencySymbol
-} from '../../../../../../libs/ui-data/src/lib/types';
-import {
-  fxEntriesToXpointData,
-  lineChartOptions
-} from '../../../../../../libs/ui-data/src/lib/functions';
-import { HISTORICAL_CHART_NAME } from 'libs/ui-data/src';
+} from '@fx/ui-core-data';
 
 interface HistoricalRatesChartProps {
   historicalRates: FxEntries;
@@ -25,12 +23,7 @@ export const HistoricalRatesChart = ({
   ]);
   return (
     <HighchartsWrapper
-      options={lineChartOptions({
-        title: `${base}/${quote}`,
-        data,
-        name: HISTORICAL_CHART_NAME,
-        stepFactor: 10
-      })}
+      options={lineChartOptions(historicalChartOptions(data, base, quote))}
     />
   );
 };
