--- libs/ui-data/.eslintrc
@@ -1,1 +1,0 @@
-{ "extends": "../../.eslintrc", "rules": {}, "ignorePatterns": ["!**/*"] }
