--- libs/ui-data/README.md
@@ -1,9 +1,0 @@
-# ui-data
-
-This library was generated with [Nx](https://nx.dev).
-
-It is a client for, FxTransaction, an [IndexedDb](https://developer.mozilla.org/en-US/docs/Web/API/IndexedDB_API) used by the demo UI apps in the workspace.
-
-## Running unit tests
-
-Run `ng test fx-transaction-db-client` to execute the unit tests via [Jest](https://jestjs.io).
