--- libs/ui-data/jest.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  name: 'fx-transaction-db-client',
-  preset: '../../jest.config.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest'
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-  coverageDirectory: '../../coverage/libs/fx-transaction-db-client'
-};
