--- libs/ui-data/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/constants';
-export * from './lib/functions';
-export * from './lib/types';
